Refactoring combine notes
Moves spendPostTime to utils
Moves fetchNotes to utils
Moves TransactionTimer to its own file

This work is part of the work to add a timer to the transaction send command.
